move api_tokens to admin module
