+++ app/components/profile/api_token_card_component/api_token_card_component.html.haml
@@ -4,7 +4,7 @@
   %p
     = t('.first_paragraph_html', application_name: APPLICATION_NAME, api_doc_url: API_DOC_URL)
 
-  = link_to t('.create_token'), nom_api_tokens_path, class: "fr-btn fr-btn--secondary fr-mt-2w"
+  = link_to t('.create_token'), nom_admin_api_tokens_path, class: "fr-btn fr-btn--secondary fr-mt-2w"
 
   %ul.fr-mt-4w
     =  render Profile::APITokenComponent.with_collection(api_tokens)
