+++ app/components/profile/api_token_component/api_token_component.html.haml
@@ -10,7 +10,7 @@
     %div= use_and_expiration
   %div
     = link_to 'Supprimer',
-      api_token_path(@api_token),
+      admin_api_token_path(@api_token),
       method: :delete,
       class: 'fr-btn fr-btn--tertiary-no-outline fr-btn--sm fr-btn--icon-left fr-icon-delete-line',
       data: { confirm: "Confirmez-vous la suppression du jeton « #{@api_token.name} » ?" }
