+++ app/views/administrateurs/api_tokens/create.html.haml
@@ -3,7 +3,7 @@
 = render partial: 'administrateurs/breadcrumbs',
   locals: { steps: [['Tableau de bord', tableau_de_bord_helper_path],
     [t('users.profil.show.profile'), profil_path],
-    [t('api_tokens.nom.new_token')]] }
+    [t('administrateurs.api_tokens.nom.new_token')]] }
 
 .fr-container.fr-mt-2w{ 'data-turbo': 'true' }
   %h1 Votre jeton est prêt
