+++ app/views/administrateurs/api_tokens/index.turbo_stream.haml
