+++ app/views/administrateurs/api_tokens/securite.html.haml
@@ -3,7 +3,7 @@
 = render partial: 'administrateurs/breadcrumbs',
   locals: { steps: [['Tableau de bord', tableau_de_bord_helper_path],
     [t('users.profil.show.profile'), profil_path],
-    [t('api_tokens.nom.new_token')]] }
+    [t('administrateurs.api_tokens.nom.new_token')]] }
 
 .fr-container.fr-mt-2w
   %h1 Sécurité
@@ -21,7 +21,7 @@
     %b Il est de votre responsabilité de le conserver en sécurité et d'en limiter l'utilisation aux seules personnes habilitées.
   %p Pour vous aider, nous vous proposons des fonctionnalités de filtrage réseau et de durée de vie du jeton.
 
-  = form_with url: api_tokens_path,
+  = form_with url: admin_api_tokens_path,
     method: :post,
     html: { class: 'fr-mt-2w' },
     data: { controller: 'api-token-securite' } do |f|
@@ -102,4 +102,4 @@
           'data-api-token-securite-target': 'continueButton' do
           créer le jeton
       %li
-        = link_to 'retour', autorisations_api_tokens_path(name: params[:name], access: params[:access], target: params[:target], targets: params[:targets]), class: "fr-btn fr-btn--secondary"
+        = link_to 'retour', autorisations_admin_api_tokens_path(name: params[:name], access: params[:access], target: params[:target], targets: params[:targets]), class: "fr-btn fr-btn--secondary"
