+++ config/locales/views/api_tokens.fr.yml
@@ -1,12 +1,13 @@
 ---
 fr:
-  api_tokens:
-    nom:
-      new_token: Création d'un nouveau jeton
-      name: Nom du jeton
-      name-hint: 'exemples: prod orus, test presta'
-      continue: Continuer
-      cancel: Retour
-    autorisations:
-      cancel: Retour
-      continue: Continuer
+  administrateurs:
+    api_tokens:
+      nom:
+        new_token: Création d'un nouveau jeton
+        name: Nom du jeton
+        name-hint: 'exemples: prod orus, test presta'
+        continue: Continuer
+        cancel: Retour
+      autorisations:
+        cancel: Retour
+        continue: Continuer
